+++ src/schema/2.0/parameter-object/query-parameter-object/array-query-parameter-object.ts
@@ -1,0 +1,17 @@
+import * as t from 'io-ts';
+import { BaseQueryParameterObjectProps } from './base-query-parameter-object';
+import { NonArrayItemsObject } from '../../items-object';
+
+export interface ArrayQueryParameterObject extends BaseQueryParameterObjectProps {
+	readonly type: 'array';
+	readonly items: NonArrayItemsObject;
+}
+
+export const ArrayQueryParameterObject = t.type(
+	{
+		...BaseQueryParameterObjectProps,
+		type: t.literal('array'),
+		items: NonArrayItemsObject,
+	},
+	'ArrayQueryParameterObject',
+);
