+++ src/schema/2.0/parameter-object/query-parameter-object/integer-query-parameter-object.ts
@@ -1,0 +1,14 @@
+import { BaseQueryParameterObjectProps } from './base-query-parameter-object';
+import * as t from 'io-ts';
+
+export interface IntegerQueryParameterObject extends BaseQueryParameterObjectProps {
+	readonly type: 'integer';
+}
+
+export const IntegerQueryParameterObject = t.type(
+	{
+		...BaseQueryParameterObjectProps,
+		type: t.literal('integer'),
+	},
+	'IntegerQueryParameterObject',
+);
