+++ src/schema/2.0/parameter-object/query-parameter-object/number-query-parameter-object.ts
@@ -1,0 +1,14 @@
+import { BaseQueryParameterObjectProps } from './base-query-parameter-object';
+import * as t from 'io-ts';
+
+export interface NumberQueryParameterObject extends BaseQueryParameterObjectProps {
+	readonly type: 'number';
+}
+
+export const NumberQueryParameterObject = t.type(
+	{
+		...BaseQueryParameterObjectProps,
+		type: t.literal('number'),
+	},
+	'NumberQueryParameterObject',
+);
