+++ src/schema/2.0/parameter-object/query-parameter-object/string-query-parameter-object.ts
@@ -1,0 +1,14 @@
+import { BaseQueryParameterObjectProps } from './base-query-parameter-object';
+import * as t from 'io-ts';
+
+export interface StringQueryParameterObject extends BaseQueryParameterObjectProps {
+	readonly type: 'string';
+}
+
+export const StringQueryParameterObject = t.type(
+	{
+		...BaseQueryParameterObjectProps,
+		type: t.literal('string'),
+	},
+	'StringQueryParameterObject',
+);
