+++ src/schema/2.0/reference-object.ts
@@ -4,6 +4,8 @@ export interface ReferenceObject {
 	readonly $ref: string;
 }
 
+export type Reference<A> = A | ReferenceObject;
+
 export const ReferenceObject = t.type(
 	{
 		$ref: t.string,
