+++ src/schema/2.0/responses-definitions-object.ts
@@ -1,7 +1,7 @@
 import { Dictionary } from '../../utils/types';
 import { ResponseObject } from './response-object';
-import * as t from 'io-ts';
+import { dictionary } from '../../utils/io-ts';
 
 export interface ResponsesDefinitionsObject extends Dictionary<ResponseObject> {}
 
-export const ResponsesDefinitionsObject = t.record(t.string, ResponseObject, 'ResponsesDefinitionsObject');
+export const ResponsesDefinitionsObject = dictionary(ResponseObject, 'ResponsesDefinitionsObject');
