+++ src/schema/2.0/scopes-object.ts
@@ -1,6 +1,7 @@
 import { Dictionary } from '../../utils/types';
-import * as t from 'io-ts';
+import { dictionary } from '../../utils/io-ts';
+import { string } from 'io-ts';
 
 export interface ScopesObject extends Dictionary<string> {}
 
-export const ScopesObject = t.record(t.string, t.string, 'ScopesObject');
+export const ScopesObject = dictionary(string, 'ScopesObject');
